Rewrite lightrec_init_mmap to not use MAP_FIXED or MAP_FIXED_NO_REPLACE
Using MAP_FIXED is likely to cause issues by clobbering existing mappings
so instead check for mmap not returning the hinted address, and retry on a
different base instead.

MAP_FIXED_NOREPLACE might not be supported on the running kernel and so
need to check returned address anyways, so not worth using.